TTS (#4)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

Прикручиваем ТТС с Парадайза.

Каждая каста ксеноморфов со своим голосом. Анноунсеры также с ТТСом.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->

https://github.com/user-attachments/assets/5cf86011-d57a-4c1f-848b-9fbeadb623f8

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
label your changes in the changelog. Please note that maintainers freely
reserve the right to remove and add tags should they deem it
appropriate. You can attempt to finagle the system all you want, but
it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Maintainers freely reserve the right to remove and add
tags should they deem it appropriate. -->

:cl:
add: Добавлен ТТС
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->