Button names on mouseover (#32)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Добавляет названия кнопкам и поддержку описания этим кнопкам при
наводке.


![image](https://github.com/user-attachments/assets/8353c538-d760-47fb-b292-40eaa29ed283)

## Summary by Sourcery

New Features:
- Display the name and description of action buttons as tooltips on
mouse hover.