runner translate (#48)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

;ЗА УЛЕЙ!!!

## Summary by Sourcery

Translate runner caste strings to Russian.

New Features:
- Add Russian translation for runner caste's abilities.

Documentation:
- Translate runner caste descriptions to Russian.