Xeno stuff (#54)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

fixes https://github.com/ss220club/BandaMarines/issues/52

Чинит икон стейт ови
Выбор касты полностью на русском языке

## Summary by Sourcery

Translate caste selection to Russian and fix the strain name to default
to "Normal" instead of "Обычный".

Bug Fixes:
- Fixed the default strain name.

Enhancements:
- Translated the caste selection menu to Russian.